Added documentation, scripts renamed
